Hotfix: Prevent app from re-creating new game
There was a nasty bug when moving the app window to another display.

Notably, Windows, upon the change in DPI, told the app to recompose itself entirely.

This meant the app threw away its GameLoop, which was defined in @Composable App() (oversight on my part).

GameLoop is now defined and initialized outside the @Composable scope entirely, and then collected as MutableState by @Composable App().

TODO: Change this to launch the main menu instead.